Did some spring cleaning and modularized the classes within the Program.cs into their own individual items. Put all in the same namespace.
