+++ Reducto_and_Absurdum/ProductCategories.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Reducto_and_Absurdum
+{
+    class ProductCategory
+    {
+        public string Name { get; set; }
+        public string Id { get; set; }
+
+    }
+}
